+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * All `FlaneurCollectionViewDelegate` functions are now optional for `UIViewController` instances
   that benefit from a default implementation
 * Added `userInfo` to `FlaneurCollectionFilter` #11
+* Added `LayoutBorderManager`: a `NSLayoutConstraint` helper #14
 
 ### Changed